anadiendo pwa app

+++ webapp/src/main/java/com/proyect/webapp/interfaces/IOva.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface IOva {
     public List<Ova>listar();
-    public Optional<Ova> listarid(int id);
+    public Optional<Ova> listarid(Long id);
     public int save(Ova o);
-    public void delete(int id);
+    public void delete(Long id);
 }
